myproject/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-appconf==1.0.6
django-imagekit==5.0.0
django-redis==5.4.0
ExifRead==3.0.0
idna==3.6
mysqlclient==2.2.4
piexif==1.1.3
pilkit==3.0
pillow==10.3.0
pip==23.3.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
setuptools==68.2.2
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.41.2
uwsgi==2.0.26
